+++ tests/test_core_io_raw.py
@@ -2,9 +2,8 @@
 """Test pycosio._core.io_raw"""
 import io
 import os
-import pytest
 
-from tests.utilities import BYTE
+import pytest
 
 
 def test_object_raw_base_io():
@@ -52,15 +51,24 @@ def test_object_raw_base_io():
         """Dummy IO"""
         _SYSTEM_CLASS = DummySystem
 
-        def _flush(self):
+        def _flush(self, buffer, *_):
             """Flush in a buffer"""
-            flushed[:] = self._write_buffer
+            flushed[:] = buffer
 
         def _read_range(self, start, end=0):
             """Read fake bytes"""
             if end == 0:
                 end = size
-            return ((size if end > size else end) - start) * BYTE
+            return ((size if end > size else end) - start) * b'0'
+
+    class DummyIORandomWrite(DummyIO):
+        """Dummy IO with random write support"""
+
+        _SUPPORT_RANDOM_WRITE = True
+
+        def _flush(self, buffer, start, stop):
+            """Flush in a buffer"""
+            flushed[start:stop] = buffer
 
     # Test seek/tell
     object_io = DummyIO(name)
@@ -85,9 +93,9 @@ def test_object_raw_base_io():
 
     # Test readinto
     object_io = DummyIO(name)
-    buffer = bytearray(100)
-    assert object_io.readinto(buffer) == 100
-    assert bytes(buffer) == 100 * BYTE
+    read_buffer = bytearray(100)
+    assert object_io.readinto(read_buffer) == 100
+    assert bytes(read_buffer) == 100 * b'0'
     assert object_io.tell() == 100
     assert object_io.readinto(bytearray(100)) == 100
     assert object_io.tell() == 200
@@ -95,17 +103,17 @@ def test_object_raw_base_io():
 
     # Test read with size (call readinto)
     object_io.seek(200)
-    assert object_io.read(100) == 100 * BYTE
+    assert object_io.read(100) == 100 * b'0'
     assert object_io.tell() == 300
 
     # Test readall
     object_io.seek(300)
-    assert object_io.readall() == (size - 300) * BYTE
+    assert object_io.readall() == (size - 300) * b'0'
     assert object_io.tell() == size
 
     # Test read without size (call readall)
     object_io.seek(300)
-    assert object_io.read() == (size - 300) * BYTE
+    assert object_io.read() == (size - 300) * b'0'
     assert object_io.tell() == size
 
     # Tests: flush should do nothing
@@ -115,18 +123,19 @@ def test_object_raw_base_io():
 
     # Test write in read mode
     with pytest.raises(io.UnsupportedOperation):
-        object_io.write(BYTE)
+        object_io.write(b'0')
 
     # Test write
     object_io = DummyIO(name, mode='w')
-    assert object_io.write(10 * BYTE) == 10
+    assert not object_io._SUPPORT_RANDOM_WRITE
+    assert object_io.write(10 * b'0') == 10
     assert object_io.tell() == 10
-    assert object_io.write(10 * BYTE) == 10
+    assert object_io.write(10 * b'0') == 10
     assert object_io.tell() == 20
     object_io.seek(10)
-    assert object_io.write(10 * BYTE) == 10
+    assert object_io.write(10 * b'0') == 10
     assert object_io.tell() == 20
-    assert object_io._get_buffer().tobytes() == 20 * BYTE
+    assert object_io._get_buffer().tobytes() == 20 * b'0'
 
     # Test flush
     assert not len(flushed)
@@ -136,7 +145,7 @@ def test_object_raw_base_io():
     # Test append
     object_io = DummyIO(name, mode='a')
     assert object_io.tell() == size
-    assert bytes(object_io._write_buffer) == size * BYTE
+    assert bytes(object_io._write_buffer) == size * b'0'
 
     # Test exclusive creation
     with pytest.raises(OSError):
@@ -148,3 +157,19 @@ def test_object_raw_base_io():
     # Test HTTP range
     assert ObjectRawIOBase._http_range(10, 50) == 'bytes=10-49'
     assert ObjectRawIOBase._http_range(10) == 'bytes=10-'
+
+    # Random write access
+    object_io = DummyIORandomWrite(name, mode='w')
+    flushed[:] = b''
+    assert not len(flushed)
+    assert object_io._SUPPORT_RANDOM_WRITE
+    assert object_io.write(100 * b'0') == 100
+    object_io.seek(50)
+    assert len(flushed) == len(100 * b'0')
+    assert object_io.write(50 * b'1') == 50
+    object_io.flush()
+    assert flushed == 50 * b'0' + 50 * b'1'
+
+    object_io = DummyIORandomWrite(name, mode='a')
+    assert object_io.tell() == size
+    assert not len(object_io._write_buffer)